refactor: 유저 저장소 컴포넌트 리팩토링
- `UsersRepository` -> `UsersRepositoryComponent` 클래스로 명칭 변경(컴포넌트 레이어에서 사용된다는 것을 강조하기 위한 목적)
- `getOneUser`, `getOneProfile` 메서드에 도입
  - findOneXX와 달리 유저를 찾을 수 없을 때 예외를 던지는 메서드
